refactored to divide up the code and organize
