--- cmd/errorHandlers.go
@@ -1,28 +1,0 @@
-package main
-
-import (
-	"net/http"
-	"runtime/debug"
-)
-
-func (app *application) serverError(w http.ResponseWriter, r *http.Request, err error) {
-	var (
-		method = r.Method
-		uri    = r.URL.RequestURI()
-		trace  = string(debug.Stack())
-	)
-
-	app.logger.Error(err.Error(), "method", method, "uri", uri, "trace", trace)
-	http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-}
-
-func (app *application) clientError(w http.ResponseWriter, status int) {
-	statusText := http.StatusText(status)
-
-	app.logger.Error("clientError", statusText, status)
-	http.Error(w, statusText, status)
-}
-
-func (app *application) notFound(w http.ResponseWriter) {
-	app.clientError(w, http.StatusNotFound)
-}
