+++ internal/app/templates.go
@@ -1,0 +1,60 @@
+package app
+
+import (
+	"html/template"
+	"io/fs"
+	"net/http"
+	"path/filepath"
+
+	"pet-clinic.bonglee.com/ui"
+)
+
+type Alert struct {
+	Msg     string
+	MsgType string
+}
+type TemplateData struct {
+	Alert    Alert
+	FlashMsg string
+	Form     any
+}
+
+const (
+	FLASH_MSG = "FLASH_MSG"
+	PRIMARY   = "primary"
+	DANGER    = "danger"
+	SUCCESS   = "success"
+	WARNING   = "warning"
+	INFO      = "info"
+)
+
+func (app *App) NewTemplateData(r *http.Request) TemplateData {
+	return TemplateData{
+		FlashMsg: app.Session.PopString(r.Context(), FLASH_MSG),
+	}
+}
+
+func CreateTemplateCache() (map[string]*template.Template, error) {
+	cache := map[string]*template.Template{}
+
+	pages, err := fs.Glob(ui.Files, "html/pages/**/*.html")
+	if err != nil {
+		return nil, err
+	}
+
+	for _, page := range pages {
+		name := filepath.Base(page)
+
+		patterns := []string{
+			"html/base.html",
+			"html/layout/nav.html",
+			page,
+		}
+
+		ts := template.Must(template.ParseFS(ui.Files, patterns...))
+
+		cache[name] = ts
+	}
+
+	return cache, nil
+}
